services/product_service/requirements/base.txt
Django==5.2
djangorestframework
djangorestframework-simplejwt
django-filter
django-mptt
django-taggit
psycopg2-binary
python-dotenv
gunicorn
requests